+++ activestorage/lib/active_storage/engine.rb
@@ -81,6 +81,7 @@ module ActiveStorage
 
       ActiveSupport.on_load(:active_record) do
         include ActiveStorage::Attached::Model
+        include ActiveStorage::Validations
       end
     end